The Malmquist-Luenberger index

The Malmquist-Luenberger Productivity Index

The Malmquist-Luenberger index (Chung, Färe and Grosskopf, 1997, and Aparicio, Pastor and Zofío, 2013) measures the change in productivity of the observation under evaluation by comparing its relative performance with respect to reference technologies corresponding to two different time periods. Productivity change can be decomposed into efficiency change and technical change under the assumption of a constant returns to scale techncology.

In this example we compute the Malmquist-Luenberger productivity index for inputs X, good outputs Y, and bad outputs B, using directions Gx, Gy, and Gb. If not specified, default directions are Gx = :Zeros, Gy = :Observed, Gb = :Observed. If not specified, default directions are Gx = :Zeros, Gy = :Observed, Gb = :Observed.

using DataEnvelopmentAnalysis

X = Array{Float64,3}(undef, 5, 1, 2)
Y = Array{Float64,3}(undef, 5, 1, 2)
B = Array{Float64,3}(undef, 5, 1, 2)

X[:,:,1] = ones(5, 1);
X[:,:,2] = ones(5, 1);

Y[:,:,1] = [7; 5; 1; 3; 4];
Y[:,:,2] = [8; 5.5; 2; 2; 4];

B[:,:,1] = [2; 5; 3; 3; 2];
B[:,:,2] = [1; 3; 2; 4; 1];

mlprod = malmluen(X, Y, B)

Mamlmquist-Luenberger environmental DEA Model 
DMUs = 5; Inputs = 1; Outputs = 1; Bad Outputs = 1; Time periods = 2
Returns to Scale = CRS
Gx = Zeros; Gy = Observed ; Gb = Observed
Referene period = Geomean
──────────────────────────────
         ML        EC       TC
──────────────────────────────
1  1.37024   1.0       1.37024
2  1.1       0.9625    1.14286
3  1.12598   1.02717   1.09619
4  0.916246  0.826389  1.10873
5  1.2792    0.954545  1.34012
──────────────────────────────
L  = Malmquist-Luenberger Productivity Index 
EC = Efficiency Change 
TC = Technological Change

malmluen Function Documentation

DataEnvelopmentAnalysis.malmluen — Function

malmluen(X, Y, B; Gx, Gy, Gb)

Compute the Malmquist-Luenberger productivity index using data envelopment analysis for inputs X, good outputs Y, and bad outputs B, using directions Gx, Gy, and Gb.

Direction specification:

The directions Gx, Gy, and Gb can be one of the following symbols.

:Zeros: use zeros.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

Optional Arguments

refperiod=:Geomean: chooses reference period for technological change: :Base, :Comparison or :Geomean.

names: a vector of strings with the names of the decision making units.