Generalized Distance Function Models

Chavas and Cox (1999) introduced a generalized distance function efficiency measure that reescales both inputs and outputs toward the frontier technology.

\[\begin{aligned} & \underset{\delta ,\mathbf{\lambda }}{\mathop{\min }}\,\quad \quad \quad \;\ \delta \\ & \text{subject}\ \text{to} \\ & \quad \quad \quad \quad \quad \ X\mathbf{\lambda } \le \delta^{1 - \alpha} {{\mathbf{x}}_{o}} \\ & \quad \quad \quad \quad \quad \;Y\mathbf{\lambda }\ \ge {{\mathbf{y}}_{o}} / \delta^{\alpha} \\ & \quad \quad \quad \quad \quad \ \mathbf{\lambda }\ge \mathbf{0}. \end{aligned}\]

The measurement of technical efficiency assuming variable returns to scale, VRS, adds the following condition:

\[\sum\nolimits_{j=1}^{n}\lambda_j=1\]

In this example we compute the generalized distance function DEA model under variable returns to scale using $0.5$ for the value of $\alpha$:

julia> X = [5 3; 2 4; 4 2; 4 8; 7 9];

julia> Y = [7 4; 10 8; 8 10; 5 4; 3 6];

julia> deagdf(X, Y, alpha = 0.5, rts = :VRS, slack = false)
Generalized DF DEA Model 
DMUs = 5; Inputs = 2; Outputs = 2
alpha = 0.5; Returns to Scale = VRS
─────────────
   efficiency
─────────────
1     0.68185
2     1.0
3     1.0
4     0.25
5     0.36
─────────────

deagdf Function Documentation

DataEnvelopmentAnalysis.deagdf — Function

deagdf(X, Y, alpha)

Compute generalized distance function data envelopment analysis model for inputs X, outputs Y, and alpha.

Optional Arguments

alpha=0.5: alpha value.

rts=:CRS: chooses constant returns to scale. For variable returns to scale choose :VRS.

slack=true: compute input and output slacks.

Xref=X: Identifies the reference set of inputs against which the units are evaluated.

Yref=Y: Identifies the reference set of outputs against which the units are evaluated.

names: a vector of strings with the names of the decision making units.

Examples

julia> X = [5 3; 2 4; 4 2; 4 8; 7 9];

julia> Y = [7 4; 10 8; 8 10; 5 4; 3 6];

julia> deagdf(X, Y, alpha = 0.5, rts = :VRS)
Generalized DF DEA Model 
DMUs = 5; Inputs = 2; Outputs = 2
alpha = 0.5; Returns to Scale = VRS
─────────────────────────────────────────────────────────────
   efficiency     slackX1     slackX2     slackY1     slackY2
─────────────────────────────────────────────────────────────
1     0.68185  0.605935    4.26672e-8  3.91163e-8  4.67865
2     1.0      5.34772e-8  6.70059e-8  2.7034e-8   4.8232e-8
3     1.0      4.82929e-8  7.26916e-8  6.00225e-8  1.66806e-8
4     0.25     4.6491e-8   9.94558e-8  5.39305e-8  9.75587e-8
5     0.36     0.2         3.4         3.0         8.07052e-8
─────────────────────────────────────────────────────────────