Hölder Distance Function Models

Briec (1998) defined technical inefficiency using Hölder norms.

Hölder L1

In this example we compute the Hölder L1 DEA model under varible returns to scale:

using DataEnvelopmentAnalysis

X = [2; 4; 8; 12; 6; 14; 14; 9.412];

Y = [1; 5; 8; 9; 3; 7; 9; 2.353];

deaholder(X, Y, l = 1, rts = :VRS)

Hölder L1 DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Orientation = Graph; Returns to Scale = VRS
────────────────────────────────────────────────
   efficiency  minimum      slackX1      slackY1
────────────────────────────────────────────────
1         0.0       X1  0.0          0.0
2         0.0       X1  0.0          0.0
3         0.0       X1  0.0          0.0
4         0.0       X1  0.0          0.0
5         3.0       X1  0.0          1.01506e-15
6         2.0       Y1  2.0          0.0
7         0.0       Y1  2.0          0.0
8         6.0       Y1  1.77636e-15  0.0
────────────────────────────────────────────────

Estimated efficiency scores are returned with the efficiency function:

holderl1 = deaholder(X, Y, l = 1, rts = :VRS);

Hölder L1 DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Orientation = Graph; Returns to Scale = VRS
────────────────────────────────────────────────
   efficiency  minimum      slackX1      slackY1
────────────────────────────────────────────────
1         0.0       X1  0.0          0.0
2         0.0       X1  0.0          0.0
3         0.0       X1  0.0          0.0
4         0.0       X1  0.0          0.0
5         3.0       X1  0.0          1.01506e-15
6         2.0       Y1  2.0          0.0
7         0.0       Y1  2.0          0.0
8         6.0       Y1  1.77636e-15  0.0
────────────────────────────────────────────────

efficiency(holderl1)

8-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 2.9999999999999996
 2.000000000000001
 0.0
 6.0

The input or output that determines the projection to the frontier is returned with:

efficiency(holderl1, :min)

8-element Vector{Int64}:
 1
 1
 1
 1
 1
 2
 2
 2

with inputs and outputs numbered sequentially.

Hölder L2

Requieres a solver that supports SOS constraints

The Hölder L2 model requieres a solver that supports SOS constraints, such as Gurobi.

Solving the model with Ipopt will return invalid results.

Hölder LInf

In this example we compute the Hölder LInf DEA model under varible returns to scale:

X = [2; 4; 8; 12; 6; 14; 14; 9.412];

Y = [1; 5; 8; 9; 3; 7; 9; 2.353];

deaholder(X, Y, l = Inf, rts = :VRS)

Hölder LInf DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Orientation = Graph; Returns to Scale = VRS
────────────────────────────────────────
   efficiency      slackX1       slackY1
────────────────────────────────────────
1       0.0    0.0           0.0
2       0.0    0.0           0.0
3       0.0    0.0           0.0
4       0.0    0.0           0.0
5       2.0    0.0          -8.08877e-16
6       2.0    0.0           0.0
7       0.0    2.0           0.0
8       3.832  2.96059e-16   0.0
────────────────────────────────────────

deaholder Function Documentation

DataEnvelopmentAnalysis.deaholder — Function

deaholder(X, Y; l)

Compute the Hölder distance function model using data envelopment analysis for inputs X and outputs Y, using Hölder norm l.

Hölder norm l specification

1.

2.

Inf.

Optional Arguments

weigt=false: set to true for weighted (weakly) Hölder distance function.

orient=:Input: chooses the radially oriented input mode. For the radially oriented output model choose :Output.

rts=:CRS: chooses constant returns to scale. For variable returns to scale choose :VRS.

slack=true: computes input and output slacks.

Xref=X: Identifies the reference set of inputs against which the units are evaluated.

Yref=Y: Identifies the reference set of outputs against which the units are evaluated.

names: a vector of strings with the names of the decision making units.