Modified Directional Distance Function

Based on the data matrix $(X,Y)$, we calculate the modified directional distance function MDDF, (Aparicio et al. 2013), of each observation o by solving $n$ times the following linear programming problem:

\[\begin{aligned} & \underset{\beta^x, \beta^y,\lambda_j }{\mathop{\min }}\,\quad \quad \quad \;\ \beta^x + \beta^y \\ & \text{subject}\ \text{to} \\ & \quad \quad \quad \quad \quad \ \sum_{j=1}^{n}{\lambda_j x_{ij} }\ \le x_{io} - \beta^x {g}_{io}^{-} \qquad i = 1,...,m \\ & \quad \quad \quad \quad \quad \ \sum_{j=1}^{n}{\lambda_j y_{rj} }\ \ge y_{ro} + \beta^y {g}_{ro}^{+} \qquad r = 1,...,s \\ & \quad \quad \quad \quad \quad \ \lambda_j \ge 0 \qquad j = 1,...,n \\ & \quad \quad \quad \quad \quad \ \beta^x \ge 0 \qquad i = 1,...,m \\ & \quad \quad \quad \quad \quad \ \beta^y \ge 0 \qquad r = 1,...,s. \\ \end{aligned}\]

with the following condition when assuming variable returns to scale:

\[\sum\nolimits_{j=1}^{n}\lambda_j=1\]

In this example we compute the modified directional distance function model under variable returns to scale using ones as directions for both inputs and outputs::

julia> X = [2; 4; 8; 12; 6; 14; 14; 9.412];

julia> Y = [1; 5; 8; 9; 3; 7; 9; 2.353];

julia> deamddf(X, Y, Gx = :Ones, Gy = :Ones, rts = :VRS)
Modified DDF DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Returns to Scale = VRS
Gx = Ones; Gy = Ones
───────────────────────────────────────────────────────
   efficiency          βx          βy  slackX1  slackY1
───────────────────────────────────────────────────────
1  7.42311e-7  0.0         7.49834e-7      0.0      0.0
2  4.03943e-7  3.95267e-7  8.67603e-9      0.0      0.0
3  4.84094e-7  4.87082e-7  0.0             0.0      0.0
4  1.32755e-6  1.33672e-6  0.0             0.0      0.0
5  4.0         2.0         2.0             0.0      0.0
6  7.33333     7.33333     0.0             0.0      0.0
7  2.0         2.0         0.0             0.0      0.0
8  8.059       5.412       2.647           0.0      0.0
───────────────────────────────────────────────────────

Estimated efficiency scores are returned with the efficiency function:

julia> deamddfvrs= deamddf(X, Y, Gx = :Ones, Gy = :Ones, rts = :VRS);

julia> efficiency(deamddfvrs)
8-element Array{Float64,1}:
 7.423112595976122e-7
 4.0394299500299253e-7
 4.840938492567068e-7
 1.3275492427779482e-6
 4.000000398014331
 7.333333863503834
 2.0000013699988735
 8.059000425674004

Estimated $\beta$ on inputs and outputs are returned with the efficiency function:

julia> efficiency(deamddfvrs, :X)
8-element Array{Float64,1}:
 0.0
 3.95266969245659e-7
 4.870818366962988e-7
 1.3367167143310045e-6
 2.000000408809775
 7.333333865967322
 2.0000013799988317
 5.412000442949787

julia> efficiency(deamddfvrs, :Y)
8-element Array{Float64,1}:
 7.498338684288291e-7
 8.67602575733351e-9
 0.0
 0.0
 1.9999999892045561
 0.0
 0.0
 2.6469999827242163

deamddf Function Documentation

DataEnvelopmentAnalysis.deamddf — Function

deamddf(X, Y; Gx, Gy)

Compute data envelopment analysis modified directional distance function model for inputs X and outputs Y, using directions Gx and Gy.

Direction specification:

The directions Gx and Gy can be one of the following symbols.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

rts=:CRS: chooses constant returns to scale. For variable returns to scale choose :VRS.

slack=true: computes input and output slacks.

Xref=X: Identifies the reference set of inputs against which the units are evaluated.

Yref=Y: Identifies the reference set of outputs against which the units are evaluated.

names: a vector of strings with the names of the decision making units.

Examples

julia> X = [2; 4; 8; 12; 6; 14; 14; 9.412];

julia> Y = [1; 5; 8; 9; 3; 7; 9; 2.353] ;

julia> deamddf(X, Y, Gx = :Ones, Gy = :Ones)
Modified DDF DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Returns to Scale = CRS
Gx = Ones; Gy = Ones
──────────────────────────────────────────────────────────
    efficiency           βx           βy  slackX1  slackY1
──────────────────────────────────────────────────────────
1   1.5         9.42849e-11   1.5             0.0      0.0
2   5.26619e-7  0.0           5.36618e-7      0.0      0.0
3   2.0         6.35153e-11   2.0             0.0      0.0
4   6.0         3.92967e-11   6.0             0.0      0.0
5   4.5         4.74326e-11   4.5             0.0      0.0
6  10.5         3.39566e-11  10.5             0.0      0.0
7   8.5         3.53945e-11   8.5             0.0      0.0
8   9.412       3.64964e-11   9.412           0.0      0.0
──────────────────────────────────────────────────────────