Reverse Directional Distance Function

In this example, we compute the Reverse Directional Distance Function (Pastor et al., 2016) DEA model for the Enhanced Russell Graph associated efficiency measure under variable returns to scale:

using DataEnvelopmentAnalysis

X = [2; 4; 8; 12; 6; 14; 14; 9.412];

Y = [1; 5; 8; 9; 3; 7; 9; 2.353];

rddferg = dearddf(X, Y, :ERG, rts = :VRS)

Reverse DDF DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Orientation = Graph; Returns to Scale = VRS
Associated efficiency measure = ERG
─────────────
   efficiency
─────────────
1    0.0
2    0.0
3    0.0
4    0.0
5    0.6
6    0.52381
7    0.142857
8    0.8
─────────────

Estimated efficiency scores are returned with the efficiency function:

efficiency(rddferg)

8-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.6000000000000002
 0.523809523809524
 0.14285714285714296
 0.8

dearddf Function Documentation

DataEnvelopmentAnalysis.dearddf — Function

dearddf(X, Y, measure)

Compute data envelopment analysis reverse directional distance function (RDDF) model for inputs X, outputs Y, and efficiency measure measure.

Measure specification:

:ERG: Enhanced Russell Graph (or Slack Based Measure (SBM)).

:MDDF: Modified Directional Distance Function.

Direction specification:

For the Modified Directional Distance Function, the directions Gx and Gy can be one of the following symbols.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

Optional Arguments

orient=:Graph: choose between graph oriented :Graph, input oriented :Input, or output oriented model :Output.

rts=:CRS: choose between constant returns to scale :CRS or variable returns to scale :VRS.

atol=1e-6: tolerance for DMU to be considered efficient.

Xref=X: Identifies the reference set of inputs against which the units are evaluated.

Yref=Y: Identifies the reference set of outputs against which the units are evaluated.

names: a vector of strings with the names of the decision making units.