Bootstrap Radial Model

The bootstrap radial DEA model (Simar and Wilson, 1998) can be calculated with the deaboot function, indicating the number of bootstrap replications in the nreps parameter. The other parameters work the same as in the radial DEA model.

A random number generator can be specified in the rng parameter for reproducibility.

using DataEnvelopmentAnalysis
using StableRNGs

X = [2, 4, 3, 5, 6]
Y = [1, 2, 3, 4, 5]

ioboot = deaboot(X, Y, orient = :Input, rts = :VRS, nreps = 200, rng = StableRNG(1234567))

Bootstrap Radial DEA Model 
DMUs = 5; Inputs = 1; Outputs = 1
Orientation = Input; Returns to Scale = VRS
Bootstrap replications = 200
────────────────────────────────────────────────────────
   Reference  Corrected       Bias  Lower 95%  Upper 95%
────────────────────────────────────────────────────────
1      1.0     0.936384  0.063616    0.75          1.0
2      0.625   0.606179  0.0188208   0.535644      0.625
3      1.0     0.940476  0.0595238   0.816986      1.0
4      0.9     0.869816  0.0301841   0.706839      0.9
5      1.0     0.90737   0.0926297   0.734459      1.0
────────────────────────────────────────────────────────
Bandwidth = 0.1341

Number of bootstrap replications

The example above uses 200 bootstrap replications for illustrative purposes. In practice, at least 1000 replications are recommended.

Bias-corrected efficiency scores are returned with the efficiency function:

efficiency(ioboot)

5-element Vector{Float64}:
 0.9363840184269926
 0.6061791889779595
 0.9404761699448017
 0.8698158968555068
 0.9073703245213588

The bias, calculated as the difference between the reference efficiency score and the bias-corrected efficiency score, is returned with the bias function:

bias(ioboot)

5-element Vector{Float64}:
 0.06361598157300741
 0.018820811022040473
 0.059523830055198346
 0.030184103144493157
 0.0926296754786412

Confidence intervals at the $95\%$, or any other desired level, are calculated with the confint function:

confint(ioboot, level = 0.95)

5×2 Matrix{Float64}:
 0.75      1.0
 0.535644  0.625
 0.816986  1.0
 0.706839  0.9
 0.734459  1.0

The optimal bandwidth computed for the model is returned with the bandwidth function:

bandwidth(ioboot)

0.13410043618697765

deaboot Function Documentation

DataEnvelopmentAnalysis.deaboot — Function

deaboot(X, Y)

Compute the bootstrap radial model using data envelopment analysis for inputs X and outputs Y.

Optional Arguments

nreps=200: number of bootstrap replications.

rng=default_rng(): random number generator.

orient=:Input: chooses the radially oriented input mode. For the radially oriented output model choose :Output.

rts=:CRS: chooses constant returns to scale. For variable returns to scale choose :VRS.

Xref=X: Identifies the reference set of inputs against which the units are evaluated.

Yref=Y: Identifies the reference set of outputs against which the units are evaluated.

disposX=:Strong: chooses strong disposability of inputs. For weak disposability choose :Weak.

disposY=:Strong: chooses strong disposability of outputs. For weak disposability choose :Weak.

names: a vector of strings with the names of the decision making units.

DataEnvelopmentAnalysis.bias — Method

bias(model::BootstrapRadialDEAModel)

Return the bias from bootstrap DEA model.

StatsAPI.confint — Method

confint(model::BootstrapRadialDEAModel; level::Real=0.95)

Compute confidence intervals for efficiency scores, with confidence level level (by default 95%).

DataEnvelopmentAnalysis.bandwidth — Method

bandwidth(model::BootstrapRadialDEAModel)

Return the optimal bandwidth of a bootstrap DEA model.