Common functions for technical models

StatsAPI.nobs — Method

nbos(model::Abstract DEAModel)

Return number of observations of a DEA model.

DataEnvelopmentAnalysis.ninputs — Method

ninputs(model::AbstractDEAModel)

Return number of inputs of a DEA model.

DataEnvelopmentAnalysis.noutputs — Method

noutputs(model::AbstractDEAModel)

Return number of outputs of a DEA model.

Base.names — Method

names(model::AbstractDEAModel)

Return the names of the decision making units (DMUs)

DataEnvelopmentAnalysis.efficiency — Method

efficiency(model::AbstractTechnicalDEAModel)

Return efficiency scores of a technical DEA model.

DataEnvelopmentAnalysis.slacks — Function

slacks(model::AbstractTechnicalDEAModel, slack::Symbol)

Return slacks of a technical DEA model.

DataEnvelopmentAnalysis.targets — Method

targets(model::AbstractTechnicalDEAModel, target::Symbol)

Return targets of a technical DEA model.

DataEnvelopmentAnalysis.peers — Function

peers(model::AbstractDEAModel)

Return peers of a DEA model.

Optional Arguments

atol=1e-10: tolerance for zero values.

namesref: a vector of strings with the names of the decision making units in the reference set.

DataEnvelopmentAnalysis.peersmatrix — Function

peersmatrix(model::AbstractDEAModel)

Return peers matrix of a DEA model.

DataEnvelopmentAnalysis.ispeer — Function

ispeer(P::DEAPeers, i::Int64, j::Int64)

Return true if j is peer of decision making unit i.

ispeer(P::DEAPeers, i::String, j::String)

Return true if j is peer of decision making unit i.

ispeer(P::DEAPeersDMU, j::Int64)

Return true if j is peer.

ispeer(P::DEAPeers, j::String)

Return true if j is peer.

DataEnvelopmentAnalysis.multipliers — Function

multipliers(model::AbstractTechnicalDEAModel, multiplier::Symbol)

Return multipliers (shadow prices) of a technical DEA model.

DataEnvelopmentAnalysis.rts — Function

rts(model::AbstractTechnicalDEAModel)

Return the value measuring the returns to scale of a multiplier DEA model.