Common functions for technical models

StatsBase.nobs — Method

nbos(model::Abstract DEAModel)

Return number of observations of a DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> nobs(deaio)
11

DataEnvelopmentAnalysis.ninputs — Method

ninputs(model::AbstractDEAModel)

Return number of inputs of a DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> ninputs(deaio)
2

DataEnvelopmentAnalysis.noutputs — Method

noutputs(model::AbstractDEAModel)

Return number of outputs of a DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> noutputs(deaio)
1

Base.names — Method

names(model::AbstractDEAModel)

Return the names of the decision making units (DMUs)

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> dmunames = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K"]

julia> deaio = dea(X, Y, names = dmunames);

julia> names(deaio)
11-element Vector{String}:
 "A"
 "B"
 "C"
 "D"
 ⋮
 "H"
 "I"
 "J"
 "K"

DataEnvelopmentAnalysis.efficiency — Method

efficiency(model::AbstractTechnicalDEAModel)

Return efficiency scores of a technical DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> efficiency(deaio)
11-element Vector{Float64}:
 1.0
 0.6222896790980051
 0.8198562443845464
 1.0
 ⋮
 0.7576690895651103
 0.8201058201058201
 0.49056603773584917
 1.0

DataEnvelopmentAnalysis.slacks — Function

slacks(model::AbstractTechnicalDEAModel, slack::Symbol)

Return slacks of a technical DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> slacks(deaio, :X)
11×2 Matrix{Float64}:
  0.0          0.0
 -4.41868e-15  0.0
  0.0          8.17926e-15
 -8.03397e-16  0.0
  ⋮            
  1.60679e-15  0.0
  1.64021      0.0
  9.68683e-15  0.0
  0.0          4.0

julia> slacks(deaio, :Y)
11×1 Matrix{Float64}:
 0.0
 0.0
 0.0
 0.0
 ⋮
 0.0
 0.0
 0.0
 0.0

DataEnvelopmentAnalysis.targets — Method

targets(model::AbstractTechnicalDEAModel, target::Symbol)

Return targets of a technical DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> targets(deaio, :X)
11×2 Matrix{Float64}:
  5.0      13.0
  9.95663   7.46748
 13.1177   21.3163
 17.0      15.0
  ⋮        
 20.4571   16.6687
 28.7037   11.4815
 20.6038   12.2642
  5.0      13.0

julia> targets(deaio, :Y)
11×1 Matrix{Float64}:
 12.0
 14.0
 25.0
 26.0
  ⋮
 30.0
 31.0
 26.0
 12.0

DataEnvelopmentAnalysis.peers — Function

peers(model::AbstractDEAModel)

Return peers of a DEA model.

Optional Arguments

atol=1e-10: tolerance for zero values.

namesref: a vector of strings with the names of the decision making units in the reference set.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> peers(deaio)
DEA Peers
1: 1 ( 1.0 ) 
2: 4 ( 0.424978317432784 ) 7 ( 0.10928013876843023 ) 
3: 1 ( 1.134321653189578 ) 4 ( 0.43800539083557943 ) 
4: 4 ( 1.0 ) 
5: 4 ( 0.25738077214231636 ) 7 ( 0.04844814534443607 ) 
6: 7 ( 0.3333333333333333 ) 
7: 7 ( 1.0 ) 
8: 4 ( 1.0348650979425895 ) 7 ( 0.11457435012935832 ) 
9: 7 ( 1.1481481481481481 ) 
10: 4 ( 0.49056603773584906 ) 7 ( 0.4905660377358491 ) 
11: 11 ( 1.0 ) 

DataEnvelopmentAnalysis.peersmatrix — Function

peersmatrix(model::AbstractDEAModel)

Return peers matrix of a DEA model.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> deaio = dea(X, Y);

julia> peersmatrix(deaio)
11×11 SparseMatrixCSC{Float64, Int64} with 17 stored entries:
 1.0       ⋅    ⋅    ⋅         ⋅    ⋅    ⋅            ⋅    ⋅    ⋅    ⋅ 
  ⋅        ⋅    ⋅   0.424978   ⋅    ⋅   0.10928       ⋅    ⋅    ⋅    ⋅ 
 1.13432   ⋅    ⋅   0.438005   ⋅    ⋅    ⋅            ⋅    ⋅    ⋅    ⋅ 
  ⋅        ⋅    ⋅   1.0        ⋅    ⋅   1.48479e-16   ⋅    ⋅    ⋅    ⋅ 
 ⋮                                 ⋮                                ⋮
  ⋅        ⋅    ⋅   1.03487    ⋅    ⋅   0.114574      ⋅    ⋅    ⋅    ⋅ 
  ⋅        ⋅    ⋅    ⋅         ⋅    ⋅   1.14815       ⋅    ⋅    ⋅    ⋅ 
  ⋅        ⋅    ⋅   0.490566   ⋅    ⋅   0.490566      ⋅    ⋅    ⋅    ⋅ 
  ⋅        ⋅    ⋅    ⋅         ⋅    ⋅    ⋅            ⋅    ⋅    ⋅   1.0

DataEnvelopmentAnalysis.ispeer — Function

ispeer(P::DEAPeers, i::Int64, j::Int64)

Return true if j is peer of decision making unit i.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> P = peers(dea(X, Y));

julia> ispeer(P, 2, 4)
true

ispeer(P::DEAPeers, i::String, j::String)

Return true if j is peer of decision making unit i.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> firms = ["A"; "B"; "C"; "D"; "E"; "F"; "G"; "H"; "I"; "J"; "K"]

julia> P = peers(dea(X, Y, names = firms));

julia> ispeer(P, "B", "D")
true

ispeer(P::DEAPeersDMU, j::Int64)

Return true if j is peer.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> P = peers(dea(X, Y));

julia> P2 = P[2];

julia> ispeer(P2, 4)
true

ispeer(P::DEAPeers, j::String)

Return true if j is peer.

Examples

julia> X = [5 13; 16 12; 16 26; 17 15; 18 14; 23 6; 25 10; 27 22; 37 14; 42 25; 5 17];

julia> Y = [12; 14; 25; 26; 8; 9; 27; 30; 31; 26; 12];

julia> firms = ["A"; "B"; "C"; "D"; "E"; "F"; "G"; "H"; "I"; "J"; "K"]

julia> P = peers(dea(X, Y, names = firms));

julia> P2 = P[2]

julia> ispeer(P2, "D")
true